Merged PR 9714: Add webapp
Ports over the backend of a webapp into the repo, including the user management, and email services.

This does not include any frontend logic, as that will come in future PRs with NextJS.

Note - we're using the .NET 7 for the webapp, but the Function app is currently 6, this will be updated.

Also:
- Makes the initial data migrations for the project.
- Restructures the repo to mirror other projects.

Related work items: #114244