Merged PR 9765: Add duplicate checks to the functions app.
Adds guards to prevent adding reports to the database, if they already exist.
If reports in the DB no longer exist in the latest redCap fetch, we update their status to resolved.
If they don't exist in the db, then they are new reports and are added.

Includes:
- Adds / seeds a ReportStatus entity ("Resolved"/"Viewed"/"Active") for now.
- Adds two step migrations for this entity (nullable, then not)
- Update the "mismatching" methods in the function app, to be called "Conflicting" now.
- Add an in memory database test fixture
- Add tests for the report service.

Related work items: #119013, #119036