Merged PR 9779: Update Reports data model
Updates the Report Data model, to implement a strongly typed list of Sites. This simplifies the data model, and instead of passing around tuples of Reports to represent a problem, one Report represents a problem, which can consist of multiple sites.

This makes the model much easier to work with and test, and means we can set up for multiple RedCap instances rather than just two.

Includes:
- Entity changes
- Data migrations
- Update service layers in function app and web app.
- Update tests

Related work items: #119904, #119910