Merged PR 9846: Add deployment pipeline
Pipeline files for a UAT and production deployment, including:

- .NET Core Webapp
- Nextjs Node App
- .NET Functions App

And new pipeline environments: `NUH RedCap Monitor UAT` + `NUH RedCap Monitor Prod`, checks needed to be added on the production environment by an admin.

Also:
- Updates an incorrect typescript import
- Removes Cors Framework package.
- Removes the npm workspaces for now, as deploying with this is not easy, see: https://github.com/npm/rfcs/issues/287
Hopefully can come back to implementing it with workspaces, but this is the most direct route to urgent deployment.
- Adds a `server.js` file for the app service to run the Next app as, via PM2: https://learn.microsoft.com/en-us/azure/app-service/configure-language-nodejs?pivots=platform-linux#run-with-pm2

Related work items: #125090, #125102