Merged PR 9867: Add Keycloak to backend
Adds Keycloak as the identity provider for the .NET backend. As in other .NET apps, we use claims to build policies, but these are now mapped from Keycloak roles.

Includes:
- Add Keycloak docker-compose as quickstart
- Add Keycloak dev and UAT realms for importing
- Update documentation
- Adds Keycloak nuget dependencies
- Removes code that is no longer necessary, or has a dependency on the .NET identity:
    - All user / account controllers, cookie auth, and claims builders.
    - Account email sending
    - The role / user management that was ported previously
- Adds authorization to the Reports endpoint, that is built with a Keycloak realm role.

I have not included removing the related entities and migrations yet - this PR is big enough so that can follow.

This is not deployable yet - we need a working Keycloak instance in Azure first (I think Owner permission is necessary to set this up).

Related work items: #120465