Merged PR 9897: Fix panda imports
Fixes the frontend app build when using shadow-panda.

The main fix here is in `panda.config.ts` - changing `emitPackage` to `false`, this isn't necessary and doesn't seem to work as expected locally in or in CI.
The imports for the ui components are updated to use the panda generated output also. Also prettier formatting.

So when copying components from shadow-panda, we need to change the imports to use the output `@/styled-system/*` and not `@/shadow-panda/*`

Having turned branch policies on - this PR needs to be merged before anything else.

Related work items: #125090