Merged PR 9945: Fix function app Site Parent conflicts
Fixes the functions app, to correctly get the Site Parent Id conflicts.

Effectively when porting the code from the console app, we missed that the comparison needs to be on the Site.Id, not the Site.SiteId (these are named in RedCap).

Updated the test to be clearer as well.

Related work items: #129651