Merged PR 10155: Project fixes
Main goal here is to fix the broken migrations pipeline.

It's broken as the Keycloak api library needed options to configure it, which don't exist when we run the migrations bundle.

So the fix is to use the new startup pattern from Biobanks/Decsys - this enables the migrations bundle to run without configuring the whole web app pipeline. Also enables us to use the CLI for anything, which is good to have.

Additionally to this:
- Remove as many warnings as possible from the build
- Update keycloak json files (had missing permissions/groups)
- Update bicep config to include missing mapping rules.

Related work items: #136075