Merged PR 10164: Add Study Capacity alerts worker
Adds a new Azure function that runs every hour, to check whether the RedCap Production studies have reached their capacity threshold, as set by the Study Groups. This is necessary, as a Study can reach its capacity without noticing.

In the backlog this is mostly referred to as Randomisation Alerts - but Study Groups / Capacity better reflects what is actually being checked, as randomisation is just one of the parts of a Study Group.

Although the job runs every hour - it will check whether the study is due to be checked, given it's requested frequency and when it was last checked.

Includes:
- New function
- Data model changes to extend a study to include capacity.
- Migrations to reflect the data model - has default values included for the existing DB.
- Tests to support this, had to fix ReportService tests to prevent a collision of using the Instances table, doesn't actually change behaviour.

Related work items: #134902, #134907