Merged PR 10335: Add shared project
Adds a project for shared code, for the functions/backend API projects.

We'd come this far without duplicating very much, but this still manages to remove a bunch of duplicated models.

But more importantly - sets us up for sharing RedCap API services, which is necessary in #141214, and generally staying away from complexity.

Related work items: #141214